+++ python/instrumentation/openinference-instrumentation-agno/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.1.23](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-agno-v0.1.22...python-openinference-instrumentation-agno-v0.1.23) (2025-11-12)
+
+
+### Bug Fixes
+
+* **agno:** Update _output_value_and_mime_type to properly deserialize output messages ([#2433](https://github.com/Arize-ai/openinference/issues/2433)) ([600d2f3](https://github.com/Arize-ai/openinference/commit/600d2f3f320bed94bcae4ae4ed956c2d178a1e1c))
+
+## [0.1.22](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-agno-v0.1.21...python-openinference-instrumentation-agno-v0.1.22) (2025-11-07)
+
+
+### Bug Fixes
+
+* Iterators for runs ([#2406](https://github.com/Arize-ai/openinference/issues/2406)) ([6ac1483](https://github.com/Arize-ai/openinference/commit/6ac14839368a932265cf64b8682586533b1633e2))
+
+## [0.1.21](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-agno-v0.1.20...python-openinference-instrumentation-agno-v0.1.21) (2025-11-06)
+
+
+### Bug Fixes
+
+* agno team id release ([#2415](https://github.com/Arize-ai/openinference/issues/2415)) ([dcd3aa7](https://github.com/Arize-ai/openinference/commit/dcd3aa7f33192a30976485c1a01b06765da24010))
+
 ## [0.1.20](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-agno-v0.1.19...python-openinference-instrumentation-agno-v0.1.20) (2025-11-03)