+++ python/instrumentation/openinference-instrumentation-haystack/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.28](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-haystack-v0.1.27...python-openinference-instrumentation-haystack-v0.1.28) (2025-11-18)
+
+
+### Features
+
+* **haystack:** Added Components instrumentation ([#2383](https://github.com/Arize-ai/openinference/issues/2383)) ([f22d546](https://github.com/Arize-ai/openinference/commit/f22d54638a655078b6919e5e05b1678b613931e3))
+
 ## [0.1.27](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-haystack-v0.1.26...python-openinference-instrumentation-haystack-v0.1.27) (2025-10-29)